+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "algebra.js": "^0.2.6",
+    "algebrite": "^1.3.1",
     "babel": "^6.23.0",
     "bootstrap": "^4.3.1",
     "expr-eval": "^2.0.2",
